Update v6.0.0
Added features:
- `-h` can also be used to display the help command
- Added FFmpeg binary within the Windows executable
- Added an exit handler
- Added 'YouTube ID' column to `songs_info.txt`

Changed features:
- Changed custom logging handler
- Changed log file path to temp folder
- Changed downloading system from titles to IDs ->
- Revamped the coding performance, up to 50x faster at times

Removed features:
- Removed 'Insert' to stop execution

+++ playlist_sync.py
@@ -3,41 +3,22 @@
 
 
 # Imports
-import main  # working with files + core
+import main
 import glob
 import os
-import mutagen.easyid3  # working with audio metadata
+import logging
 
 
 # Core
-for full_mp3_file in glob.glob('{}/*.mp3'.format(main.path_song)):  # parse through mp3 files
-    for remote_song in main.json_data['entries']:  # parse through playlist metadata
-        if remote_song:  # checks for song vaildation on YouTube
-            main.logwrite("Checking '{}' with '{}'...".format(full_mp3_file,
-                            remote_song['title']))
-            local_song = mutagen.easyid3.EasyID3(full_mp3_file)  # open mp3 file
+def sync():
+    if not glob.glob(f"{main.path_song}/*.mp3"):
+        logging.warning('Nothing to sync - no files.')
+        return
 
-            # gather local song metadata
-            local_song_data = [local_song['title'][0]]
-            if 'artist' in local_song.keys():
-                local_song_data.append(local_song['artist'])
-            if 'album' in local_song.keys():
-                local_song_data.append(local_song['album'][0])
-            
-            # gather remote song metadata
-            remote_song_data = [remote_song['title']]
-            if 'creator' in remote_song.keys():
-                remote_song_data.append(remote_song['artist'].split(', '))
-            if 'album' in remote_song.keys():
-                remote_song_data.append(remote_song['album'])
-
-            # compare the data...
-            if local_song_data == remote_song_data:
-                main.logwrite('Match found between "{}" and "{}". Continuing...'.format(local_song['title'
-                                ][0], remote_song['title']))
-                break
-    else:
-        main.logwrite("Match not found. Deleting '{}'...".format(full_mp3_file))
-        os.remove(full_mp3_file)
-else:
-    main.errorwrite('Nothing to sync - no files.', 1)
+    for full_mp3_file in glob.glob(f'{main.path_song}/*.mp3'): # parse through mp3 files
+        song_id = full_mp3_file.rsplit(os.sep, 1)[1][:full_mp3_file.rsplit(os.sep, 1)[1].find('.mp3')]
+        if song_id in main.songs_data:
+            logging.debug(f"'{song_id}' is still liked.")
+        else:
+            logging.debug(f"'{song_id}' is not liked. Removing...")
+            os.remove(full_mp3_file)
